feat: move molstar wrapper functions into lib/molstar folder
